# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a toge server API documentation.",
        "title": "toge API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/persons/": {
            "get": {
                "description": "获取所有Person列表（支持分页、排序和搜索）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "获取所有Person",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码，默认为1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "每页大小，默认为10，最大100",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段：ID, Name, Status, ",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序方向：asc, desc，默认为desc",
                        "name": "sort_order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索字段：Name, Email, ，不指定则在所有字段中搜索",
                        "name": "search_by",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_pagination.PageResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/git_lulumia_fun_root_toge-api_internal_model.Person"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的Person",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "创建Person",
                "parameters": [
                    {
                        "description": "Person信息",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_internal_domain.CreatePersonRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_internal_model.Person"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/persons/{id}": {
            "get": {
                "description": "根据ID获取Person详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "获取Person详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PersonID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_internal_model.Person"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "根据ID更新Person信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "更新Person",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PersonID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Person更新信息",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_internal_domain.UpdatePersonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_internal_model.Person"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除Person",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "persons"
                ],
                "summary": "删除Person",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PersonID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/git_lulumia_fun_root_toge-api_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "git_lulumia_fun_root_toge-api_internal_domain.CreatePersonRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "company": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "git_lulumia_fun_root_toge-api_internal_domain.UpdatePersonRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "company": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "git_lulumia_fun_root_toge-api_internal_model.Person": {
            "description": "人员信息",
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "北京市朝阳区"
                },
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "company": {
                    "type": "string",
                    "example": "科技有限公司"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "zhangsan@example.com"
                },
                "gender": {
                    "type": "string",
                    "example": "男"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "张三"
                },
                "phone": {
                    "type": "string",
                    "example": "13800138000"
                },
                "position": {
                    "type": "string",
                    "example": "软件工程师"
                },
                "status": {
                    "description": "1: 在职, 0: 离职",
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
...
        "git_lulumia_fun_root_toge-api_pkg_pagination.PageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据列表"
                },
                "has_next": {
                    "description": "是否有下一页",
                    "type": "boolean"
                },
                "has_prev": {
                    "description": "是否有上一页",
                    "type": "boolean"
                },
                "page": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "description": "总记录数",
                    "type": "integer"
                },
                "total_pages": {
                    "description": "总页数",
                    "type": "integer"
                }
            }
        },
        "git_lulumia_fun_root_toge-api_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "error": {
                    "description": "错误信息",
                    "type": "string"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}